Change LICENSE and rename package, upgrade version to 0.3.0
