+++ pom.xml
@@ -1,31 +1,38 @@
-<!-- Copyright (c) 2014 Wandoujia Inc. Permission is hereby granted, free of charge, to any person obtaining
-    a copy of this software and associated documentation files (the "Software"), to deal in the Software
-    without restriction, including without limitation the rights to use, copy, modify, merge, publish,
-    distribute,
-    sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished
-    to do so, subject to the following conditions: The above copyright notice and this permission notice
-    shall be included in all copies or substantial portions of the Software. THE SOFTWARE IS PROVIDED
-    "AS
-    IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-    OF
-    MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS
-    OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
-    CONTRACT,
-    TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-    IN THE SOFTWARE. -->
+<!--
+ Copyright (c) 2014 CodisLabs.
+
+ Permission is hereby granted, free of charge, to any person obtaining
+ a copy of this software and associated documentation files (the
+ "Software"), to deal in the Software without restriction, including
+ without limitation the rights to use, copy, modify, merge, publish,
+ distribute, sublicense, and/or sell copies of the Software, and to
+ permit persons to whom the Software is furnished to do so, subject to
+ the following conditions:
+
+ The above copyright notice and this permission notice shall be
+ included in all copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+-->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.wandoulabs.jodis</groupId>
     <artifactId>jodis</artifactId>
-    <version>0.2.2</version>
+    <version>0.3.0</version>
     <packaging>jar</packaging>
     <name>${project.artifactId}</name>
     <description>Round Robin jedis pool for codis</description>
     <url>https://github.com/wandoulabs/codis</url>
     <scm>
-        <connection>scm:git:ssh://github.com:wandoulabs/codis.git</connection>
-        <developerConnection>scm:git:ssh://github.com:wandoulabs/codis.git</developerConnection>
+        <connection>scm:git:ssh://github.com:wandoulabs/jodis.git</connection>
+        <developerConnection>scm:git:ssh://github.com:wandoulabs/jodis.git</developerConnection>
         <url>ssh://github.com:wandoulabs/codis.git</url>
         <tag>HEAD</tag>
     </scm>
