+++ src/test/java/io/codis/jodis/TestRoundRobinJedisPool.java
@@ -1,7 +1,7 @@
 /**
  * @(#)TestRoundRobinJedisPool.java, 2014-12-1. 
  * 
- * Copyright (c) 2014 Wandoujia Inc.
+ * Copyright (c) 2014 CodisLabs.
  * 
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.wandoulabs.jodis;
+package io.codis.jodis;
 
 import static org.junit.Assert.assertEquals;
 
@@ -48,6 +48,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.io.Closeables;
 
+import io.codis.jodis.RoundRobinJedisPool;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.exceptions.JedisException;
